Break out internals from language files
