+++ ts/go/consts.ts
@@ -1,0 +1,2 @@
+export const GOMOD2NIX_REPO =
+  "github:nix-community/gomod2nix?rev=f95720e89af6165c8c0aa77f180461fe786f3c21";
