+++ ts/go/mod.ts
@@ -3,15 +3,9 @@ import { NixExpression, nixRaw, nixStrLit } from "../nix.ts";
 import { mkPackage, Package } from "../package.ts";
 import { mkProject, Project } from "../project.ts";
 import * as path from "https://deno.land/std@0.202.0/path/mod.ts";
-import * as fs from "https://deno.land/std@0.201.0/fs/mod.ts";
 import { getDotGarnProjectDir } from "../internal/garn_dir.ts";
 import { nixSource } from "../internal/utils.ts";
-import outdent from "https://deno.land/x/outdent@v0.8.0/mod.ts";
-import { Initializer } from "../base.ts";
-import { camelCase } from "https://deno.land/x/case@2.2.0/mod.ts";
-
-export const GOMOD2NIX_REPO =
-  "github:nix-community/gomod2nix?rev=f95720e89af6165c8c0aa77f180461fe786f3c21";
+import { GOMOD2NIX_REPO } from "./consts.ts";
 
 const getGoModNixToml = (src: string): NixExpression => {
   const gen = new Deno.Command("nix", {
@@ -95,58 +89,3 @@ export const mkGoProject = (args: {
     }
   );
 };
-
-function parseGoMod(goModContents: string): {
-  moduleName: string;
-  goVersion: string;
-} {
-  let moduleName: string | null = null;
-  let goVersion: string | null = null;
-  for (const line of goModContents.split("\n")) {
-    const [keyword, rest] = line.split(" ");
-    switch (keyword) {
-      case "module":
-        moduleName = rest.slice(rest.lastIndexOf("/") + 1);
-        break;
-      case "go":
-        goVersion = rest;
-        break;
-    }
-    if (moduleName && goVersion) return { moduleName, goVersion };
-  }
-  throw new Error("go.mod missing module name or go version");
-}
-
-// Initializers
-const goModuleInitializer: Initializer = () => {
-  if (!fs.existsSync("go.mod")) {
-    return { tag: "ShouldNotRun" };
-  }
-  try {
-    const { goVersion, moduleName } = parseGoMod(
-      Deno.readTextFileSync("go.mod")
-    );
-    return {
-      tag: "ShouldRun",
-      makeTarget: () =>
-        outdent`
-          export const ${camelCase(moduleName)} = garn.go.mkGoProject({
-            description: "My go project",
-            moduleName: ${JSON.stringify(moduleName)},
-            src: ".",
-            goVersion: ${JSON.stringify(goVersion)},
-          });
-        `,
-    };
-  } catch (e) {
-    if (e instanceof Error) {
-      return {
-        tag: "ShouldNotRun",
-        reason: e.message,
-      };
-    }
-    throw e;
-  }
-};
-
-export const initializers = [goModuleInitializer];
